refactor: update directories :recycle:
